Select Bromeliaceae Species Ranges by Taxonomy and Traits

Get the geographic range for all species selected via the arguments. The type of range estimate depends on the “type” argument.

get_range(scientific = NULL, canonical = NULL, genus = NULL,
  subfamily = NULL, life_form = NULL, assessment = NULL,
  range_size = NULL, type = "polygon")

Arguments

scientific
a character vector of full scientific names including authorities, of the species of interest

canonical
a character vector of canonical names, of the species of interest.

genus
a character vector of genera names to select.

subfamily
a character vector of subfamily names to select.

life_form
a character vector of life forms to select.

assessment
a character vector of conservation assessment to select.

range_size
a vector of two numericals with the minimum and maximum range size (in square kilometres), to select.

type
a cahracter defining the output format, see details

Value

Depending on the “type” argument. If “binary” a presence/absence raster based on the modelled habitat suitability, at 100x100 km resolution in Behrmann projection, if “suitability” the habitat suitability as predicted by an ensemble model in Behrmann projection, and if polygon a simple feature object in lat/lon projection. .

Details

Modelled ranges are available for 542 species, range polygons for 2395 species. For species with model distribution, the range polygons are based on the models, otherwise on a convex hull around available occurrence records, or a 50 km buffer for species with less than 3 occurrence records available. See Zizka et al 2019 for methods.

Examples

get_range(scientific = "Aechmea mexicana Baker")

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -105.1333 ymin: -0.7421289 xmax: -73.88333 ymax: 23.32705
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#>        tax_accepted_name                       geometry
#> 1 Aechmea mexicana Baker MULTIPOLYGON (((-79.13333 -...

get_range(scientific = "Aechmea mexicana Baker", type = "binary")

#> $`Aechmea mexicana Baker`
#> class      : RasterLayer 
#> dimensions : 164, 162, 26568  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -10626355, -6718661, -577104.6, 3378833  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_mexicana_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> attr(,"class")
#> [1] "bromeli" "list"   

get_range(scientific = "Aechmea mexicana Baker", type = "suitability")

#> $`Aechmea mexicana Baker`
#> class      : RasterLayer 
#> dimensions : 164, 162, 26568  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -10626355, -6718661, -577104.6, 3378833  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_mexicana_Baker 
#> values     : 0.0004018875, 0.5196612  (min, max)
#> 
#> 
#> attr(,"class")
#> [1] "bromeli" "list"   

get_range(canonical = "Aechmea mexicana")

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -105.1333 ymin: -0.7421289 xmax: -73.88333 ymax: 23.32705
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#>        tax_accepted_name                       geometry
#> 1 Aechmea mexicana Baker MULTIPOLYGON (((-79.13333 -...

get_range(genus = "Aechmea")

#> Simple feature collection with 235 features and 1 field
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: -106.1333 ymin: -37.78534 xmax: -34.63333 ymax: 23.53568
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>                                               tax_accepted_name
#> 1                                    Aechmea abbreviata L.B.Sm.
#> 2                Aechmea aculeatosepala (Rauh & Barthlott) Leme
#> 3  Aechmea aenigmatica Lopez-Ferrari, Espejo, Ceja et A.Mendoza
#> 4                        Aechmea aguadocensis Leme & L.Kollmann
#> 5                                     Aechmea aiuruocensis Leme
#> 6                                              Aechmea alba Mez
#> 7                                    Aechmea alegrensis W.Weber
#> 8                                        Aechmea alopecurus Mez
#> 9                         Aechmea amicorum B.R.Silva & H.Luther
#> 10                                        Aechmea ampla L.B.Sm.
#>                          geometry
#> 1  MULTIPOLYGON (((-75.88333 -...
#> 2  POLYGON ((-78.1828 -3.26, -...
#> 3  POLYGON ((-96.4525 16.1917,...
#> 4  POLYGON ((-40.4689 -18.5706...
#> 5  POLYGON ((-44.1391 -22.0335...
#> 6  MULTIPOLYGON (((-40.13333 -...
#> 7  POLYGON ((-41 -20.7089, -41...
#> 8  POLYGON ((-43.6783 -22.5497...
#> 9  POLYGON ((-39.16467 -17.048...
#> 10 POLYGON ((-38.92074 -15.997...

get_range(genus = "Aechmea", type = "binary")

#> $`Aechmea abbreviata L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 49, 48, 2352  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7972982, -6815147, -649490.8, 532466  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_abbreviata_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea alba Mez`
#> class      : RasterLayer 
#> dimensions : 74, 53, 3922  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4547720, -3269277, -2821077, -1036081  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_alba_Mez 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea angustifolia Poepp. & Endl.`
#> class      : RasterLayer 
#> dimensions : 205, 187, 38335  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8744872, -4234139, -2628047, 2316874  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_angustifolia_Poepp._._Endl. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea aquilega (Salisb.) Griseb.`
#> class      : RasterLayer 
#> dimensions : 177, 236, 41772  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8576021, -2883331, -2410889, 1858629  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_aquilega_.Salisb.._Griseb. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea blanchetiana (Baker) L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 105, 55, 5775  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4547720, -3221033, -2965850, -433085.2  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_blanchetiana_.Baker._L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea blumenavii Reitz`
#> class      : RasterLayer 
#> dimensions : 51, 48, 2448  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5343731, -4185896, -3955128, -2724928  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_blumenavii_Reitz 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea bracteata (Sw.) Griseb.`
#> class      : RasterLayer 
#> dimensions : 112, 147, 16464  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -10674598, -7128727, 605203.6, 3306819  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_bracteata_.Sw.._Griseb. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea brevicollis L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 66, 64, 4224  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7369943, -5826163, -625362.1, 966661.4  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_brevicollis_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea bromeliifolia (Rudge) Baker`
#> class      : RasterLayer 
#> dimensions : 166, 142, 23572  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -6766904, -3341641, -4727248, -723067.5  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_bromeliifolia_.Rudge._Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea calyculata (E.Morren) Baker`
#> class      : RasterLayer 
#> dimensions : 64, 86, 5504  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -6284473, -4210018, -4172287, -2628506  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_calyculata_.E.Morren._Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea castelnavii Baker`
#> class      : RasterLayer 
#> dimensions : 143, 224, 32032  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8648386, -3245155, -1711155, 1738229  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_castelnavii_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea caudata Lindm.`
#> class      : RasterLayer 
#> dimensions : 71, 69, 4899  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5319610, -3655222, -4027514, -2314883  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_caudata_Lindm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea chantinii (Carrière) Baker`
#> class      : RasterLayer 
#> dimensions : 127, 94, 11938  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8093590, -5826163, -1831799, 1231640  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_chantinii_.Carrière._Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea coelestis (K.Koch) E.Morren`
#> class      : RasterLayer 
#> dimensions : 64, 73, 4672  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5174880, -3414006, -3544940, -2001159  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_coelestis_.K.Koch._E.Morren 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea contracta (Mart. ex Schult. & Schult.f.) Baker`
#> class      : RasterLayer 
#> dimensions : 107, 114, 12198  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7780010, -5030151, -1831799, 749208.7  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_contracta_.Mart._ex_Schult._._Schult.f.._Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea corymbosa (Mart. ex Schult. & Schult.f.) Mez`
#> class      : RasterLayer 
#> dimensions : 109, 103, 11227  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7900618, -5416096, -1614640, 1014611  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_corymbosa_.Mart._ex_Schult._._Schult.f.._Mez 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea costantinii (Mez) L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 64, 50, 3200  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4113531, -2907453, -1855928, -312147.4  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_costantinii_.Mez._L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea cylindrata Lindm.`
#> class      : RasterLayer 
#> dimensions : 77, 59, 4543  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5271367, -3848194, -3858613, -2001252  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_cylindrata_Lindm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea dactylina Baker`
#> class      : RasterLayer 
#> dimensions : 118, 76, 8968  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8672508, -6839269, -504718.4, 2341627  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_dactylina_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea distichantha Lem.`
#> class      : RasterLayer 
#> dimensions : 143, 145, 20735  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7128727, -3631100, -4727248, -1277864  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_distichantha_Lem. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea fernandae (E.Morren) Baker`
#> class      : RasterLayer 
#> dimensions : 90, 93, 8370  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -6863390, -4620084, -1083808, 1087133  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_fernandae_.E.Morren._Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea fulgens Brongn.`
#> class      : RasterLayer 
#> dimensions : 87, 58, 5046  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4330626, -2931575, -2555661, -457084.8  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_fulgens_Brongn. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea gamosepala Wittm.`
#> class      : RasterLayer 
#> dimensions : 67, 51, 3417  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5343731, -4113531, -4099901, -2483756  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_gamosepala_Wittm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea gracilis Lindm.`
#> class      : RasterLayer 
#> dimensions : 52, 58, 3016  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5223124, -3824073, -3641455, -2387133  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_gracilis_Lindm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea hoppii (Harms) L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 78, 66, 5148  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8117712, -6525688, -1252709, 628773  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_hoppii_.Harms._L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea kertesziae Reitz`
#> class      : RasterLayer 
#> dimensions : 56, 47, 2632  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5319610, -4185896, -4075772, -2724964  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_kertesziae_Reitz 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea lamarchei (Morr.) Mez`
#> class      : RasterLayer 
#> dimensions : 67, 58, 3886  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4740692, -3341641, -3351910, -1735765  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_lamarchei_.Morr.._Mez 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea leptantha (Harms) Leme & J A Siqueira`
#> class      : RasterLayer 
#> dimensions : 87, 60, 5220  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4330626, -2883331, -2410889, -312312.4  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_leptantha_.Harms._Leme_._J_A_Siqueira 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea longifolia (Rudge) L.B.Sm. & M.A.Spencer`
#> class      : RasterLayer 
#> dimensions : 162, 145, 23490  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8045347, -4547720, -2628047, 1279647  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_longifolia_.Rudge._L.B.Sm._._M.A.Spencer 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea lueddemanniana (K.Koch) Mez`
#> class      : RasterLayer 
#> dimensions : 89, 90, 8010  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -9854465, -7683523, 822362.3, 2969182  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_lueddemanniana_.K.Koch._Mez 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea maasii Gouda & W.Till`
#> class      : RasterLayer 
#> dimensions : 53, 48, 2544  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4499476, -3341641, -3134751, -1856308  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_maasii_Gouda_._W.Till 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea magdalenae (André) André ex Baker`
#> class      : RasterLayer 
#> dimensions : 159, 133, 21147  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -9782100, -6573931, -1035551, 2799779  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_magdalenae_.André._André_ex_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea marauensis Leme`
#> class      : RasterLayer 
#> dimensions : 68, 55, 3740  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4427112, -3100426, -2507404, -867137.1  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_marauensis_Leme 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea mariae-reginae H.Wendl`
#> class      : RasterLayer 
#> dimensions : 52, 49, 2548  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8696629, -7514673, 701718.6, 1956040  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_mariae.reginae_H.Wendl 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea melinonii Hook.`
#> class      : RasterLayer 
#> dimensions : 116, 110, 12760  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -6839269, -4185896, -1566383, 1231719  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_melinonii_Hook. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea mertensii (Meyer) Schult. & Schult.f.`
#> class      : RasterLayer 
#> dimensions : 164, 213, 34932  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8021225, -2883331, -2290245, 1665692  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_mertensii_.Meyer._Schult._._Schult.f. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea mexicana Baker`
#> class      : RasterLayer 
#> dimensions : 164, 162, 26568  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -10626355, -6718661, -577104.6, 3378833  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_mexicana_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea miniata (Beer) hort. ex Baker`
#> class      : RasterLayer 
#> dimensions : 58, 49, 2842  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4330626, -3148669, -2410889, -1011838  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_miniata_.Beer._hort._ex_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea multiflora L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 65, 56, 3640  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4402990, -3052182, -2435018, -867115.6  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_multiflora_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea nidularioides L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 74, 72, 5328  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7997104, -6260351, -1252709, 532286.7  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_nidularioides_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea nivea L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 55, 65, 3575  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7804131, -6236229, -673619.6, 653066.7  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_nivea_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea nudicaulis (L.) Griseb.`
#> class      : RasterLayer 
#> dimensions : 309, 289, 89301  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -9878586, -2907453, -4148158, 3305406  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_nudicaulis_.L.._Griseb. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea organensis Wawra`
#> class      : RasterLayer 
#> dimensions : 67, 70, 4690  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5223124, -3534614, -3665583, -2049438  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_organensis_Wawra 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea ornata Baker`
#> class      : RasterLayer 
#> dimensions : 61, 59, 3599  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5560825, -4137653, -3955128, -2483712  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_ornata_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea pectinata Baker`
#> class      : RasterLayer 
#> dimensions : 60, 64, 3840  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -5199002, -3655222, -3737969, -2290675  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_pectinata_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea penduliflora André`
#> class      : RasterLayer 
#> dimensions : 166, 148, 24568  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8744872, -5174880, -2121344, 1882836  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_penduliflora_André 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea phanerophlebia Baker`
#> class      : RasterLayer 
#> dimensions : 63, 57, 3591  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4764814, -3389884, -3231266, -1711607  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_phanerophlebia_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea pineliana (Brongn. ex Planch.) Baker`
#> class      : RasterLayer 
#> dimensions : 60, 55, 3300  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4740692, -3414006, -3351910, -1904616  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_pineliana_.Brongn._ex_Planch.._Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea poitaei (Baker) L.B.Sm. & M.A.Spencer`
#> class      : RasterLayer 
#> dimensions : 69, 76, 5244  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8069469, -6236229, -1132066, 532322.6  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_poitaei_.Baker._L.B.Sm._._M.A.Spencer 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea politii L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 63, 103, 6489  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7201092, -4716571, -311688.5, 1207970  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_politii_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea pubescens Baker`
#> class      : RasterLayer 
#> dimensions : 96, 101, 9696  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8937845, -6501567, 170886.3, 2486557  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_pubescens_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea pyramidalis Benth`
#> class      : RasterLayer 
#> dimensions : 63, 48, 3024  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8286563, -7128727, -987293.2, 532365.6  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_pyramidalis_Benth 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea ramosa Mart. ex Schult. & Schult.f.`
#> class      : RasterLayer 
#> dimensions : 62, 55, 3410  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4692449, -3365763, -3351910, -1856372  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_ramosa_Mart._ex_Schult._._Schult.f. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea recurvata (Klotzsch) L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 97, 106, 10282  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -6766904, -4210018, -4727248, -2387456  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_recurvata_.Klotzsch._L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea retusa L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 67, 78, 5226  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8069469, -6187986, -939035.7, 677109.4  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_retusa_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea rubiginosa Mez`
#> class      : RasterLayer 
#> dimensions : 97, 126, 12222  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8045347, -5006029, -1132066, 1207726  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_rubiginosa_Mez 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea saxicola L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 56, 47, 2632  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -4475355, -3341641, -3158880, -1808072  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_saxicola_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea servitensis André`
#> class      : RasterLayer 
#> dimensions : 114, 75, 8550  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8069469, -6260351, -939035.7, 1810823  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_servitensis_André 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea setigera Mart. ex Schult. & Schult.f.`
#> class      : RasterLayer 
#> dimensions : 166, 174, 28884  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8190076, -3992924, -2338503, 1665678  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_setigera_Mart._ex_Schult._._Schult.f. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea streptocalycoides Philcox`
#> class      : RasterLayer 
#> dimensions : 78, 59, 4602  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8069469, -6646296, -1421610, 459871.8  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_streptocalycoides_Philcox 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea strobilacea L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 157, 79, 12403  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8165955, -6260351, -2097215, 1689871  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_strobilacea_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea tessmannii Harms`
#> class      : RasterLayer 
#> dimensions : 63, 59, 3717  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8045347, -6622174, -1035551, 484108.1  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_tessmannii_Harms 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea tillandsioides (Mart. ex Schult.f.) Baker`
#> class      : RasterLayer 
#> dimensions : 206, 235, 48410  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -9830343, -4161775, -2024829, 2944214  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_tillandsioides_.Mart._ex_Schult.f.._Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea tocantina Baker`
#> class      : RasterLayer 
#> dimensions : 170, 147, 24990  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7659402, -4113531, -2652176, 1448490  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_tocantina_Baker 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea vallerandii (Carrière) Erhardt, Götz & Seybold`
#> class      : RasterLayer 
#> dimensions : 139, 153, 21267  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -7997104, -4306504, -1831799, 1521099  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_vallerandii_.Carrière._Erhardt._Götz_._Seybold 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea woronowii Harms`
#> class      : RasterLayer 
#> dimensions : 65, 70, 4550  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8093590, -6405080, -1035551, 532351.3  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_woronowii_Harms 
#> values     : 1, 1  (min, max)
#> 
#> 
#> $`Aechmea zebrina L.B.Sm.`
#> class      : RasterLayer 
#> dimensions : 47, 47, 2209  (nrow, ncol, ncell)
#> resolution : 24121.57, 24121.57  (x, y)
#> extent     : -8021225, -6887512, -673619.6, 460094.1  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : Aechmea_zebrina_L.B.Sm. 
#> values     : 1, 1  (min, max)
#> 
#> 
#> attr(,"class")
#> [1] "bromeli" "list"   

get_range(genus = c("Aechmea", "Zizkaea"))

#> Simple feature collection with 236 features and 1 field
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: -106.1333 ymin: -37.78534 xmax: -34.63333 ymax: 23.53568
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>                                               tax_accepted_name
#> 1                                    Aechmea abbreviata L.B.Sm.
#> 2                Aechmea aculeatosepala (Rauh & Barthlott) Leme
#> 3  Aechmea aenigmatica Lopez-Ferrari, Espejo, Ceja et A.Mendoza
#> 4                        Aechmea aguadocensis Leme & L.Kollmann
#> 5                                     Aechmea aiuruocensis Leme
#> 6                                              Aechmea alba Mez
#> 7                                    Aechmea alegrensis W.Weber
#> 8                                        Aechmea alopecurus Mez
#> 9                         Aechmea amicorum B.R.Silva & H.Luther
#> 10                                        Aechmea ampla L.B.Sm.
#>                          geometry
#> 1  MULTIPOLYGON (((-75.88333 -...
#> 2  POLYGON ((-78.1828 -3.26, -...
#> 3  POLYGON ((-96.4525 16.1917,...
#> 4  POLYGON ((-40.4689 -18.5706...
#> 5  POLYGON ((-44.1391 -22.0335...
#> 6  MULTIPOLYGON (((-40.13333 -...
#> 7  POLYGON ((-41 -20.7089, -41...
#> 8  POLYGON ((-43.6783 -22.5497...
#> 9  POLYGON ((-39.16467 -17.048...
#> 10 POLYGON ((-38.92074 -15.997...

get_range(subfamily = "Pitcairnioideae")

#> Simple feature collection with 607 features and 1 field
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: -109.1333 ymin: -35.86597 xmax: -35.38333 ymax: 28.36462
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>                                           tax_accepted_name
#> 1  Deuterocohnia brevifolia (Griseb.) M.A.Spencer & L.B.Sm.
#> 2                 Deuterocohnia brevispicata Rauh & L.Hrom.
#> 3                      Deuterocohnia chrysantha (Phil.) Mez
#> 4                            Deuterocohnia digitata L.B.Sm.
#> 5                 Deuterocohnia gableana R.Vásquez & Ibisch
#> 6                          Deuterocohnia glandulosa E.Gross
#> 7                            Deuterocohnia haumanii A.Cast.
#> 8                     Deuterocohnia longipetala (Baker) Mez
#> 9        Deuterocohnia lotteae (Rauh) M.A.Spencer & L.B.Sm.
#> 10                      Deuterocohnia meziana Kuntze ex Mez
#>                          geometry
#> 1  MULTIPOLYGON (((-65.38333 -...
#> 2  POLYGON ((-63.7272 -19.8536...
#> 3  MULTIPOLYGON (((-70.42 -26,...
#> 4  POLYGON ((-63.5975 -18.4167...
#> 5  POLYGON ((-63.0641 -18.1, -...
#> 6  POLYGON ((-63.133 -19.7064,...
#> 7  POLYGON ((-68.58 -30.15, -6...
#> 8  MULTIPOLYGON (((-65.13333 -...
#> 9  POLYGON ((-64.0004 -21.4894...
#> 10 MULTIPOLYGON (((-52.13333 -...

get_range(life_form = "epiphyte")

#> Simple feature collection with 1395 features and 1 field
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: -111.8833 ymin: -39.22848 xmax: -34.63333 ymax: 36.38029
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>                                               tax_accepted_name
#> 1                                    Aechmea abbreviata L.B.Sm.
#> 2                Aechmea aculeatosepala (Rauh & Barthlott) Leme
#> 3  Aechmea aenigmatica Lopez-Ferrari, Espejo, Ceja et A.Mendoza
#> 4                        Aechmea aguadocensis Leme & L.Kollmann
#> 5                                     Aechmea aiuruocensis Leme
#> 6                                    Aechmea alegrensis W.Weber
#> 7                                        Aechmea alopecurus Mez
#> 8                         Aechmea amicorum B.R.Silva & H.Luther
#> 9                                         Aechmea ampla L.B.Sm.
#> 10                         Aechmea andersonii H.H.Luther & Leme
#>                          geometry
#> 1  MULTIPOLYGON (((-75.88333 -...
#> 2  POLYGON ((-78.1828 -3.26, -...
#> 3  POLYGON ((-96.4525 16.1917,...
#> 4  POLYGON ((-40.4689 -18.5706...
#> 5  POLYGON ((-44.1391 -22.0335...
#> 6  POLYGON ((-41 -20.7089, -41...
#> 7  POLYGON ((-43.6783 -22.5497...
#> 8  POLYGON ((-39.16467 -17.048...
#> 9  POLYGON ((-38.92074 -15.997...
#> 10 MULTIPOLYGON (((-39.0244 -1...

get_range(assessment = c("CR", "VU"))

#> Simple feature collection with 1771 features and 1 field
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: -110.156 ymin: -43.3819 xmax: -34.79361 ymax: 29.9247
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>                                               tax_accepted_name
#> 1          Acanthostachys pitcairnioides (Mez) Rauh & Barthlott
#> 2                Aechmea aculeatosepala (Rauh & Barthlott) Leme
#> 3  Aechmea aenigmatica Lopez-Ferrari, Espejo, Ceja et A.Mendoza
#> 4                        Aechmea aguadocensis Leme & L.Kollmann
#> 5                                     Aechmea aiuruocensis Leme
#> 6                                    Aechmea alegrensis W.Weber
#> 7                         Aechmea amicorum B.R.Silva & H.Luther
#> 8                                         Aechmea ampla L.B.Sm.
#> 9                                       Aechmea anomala L.B.Sm.
#> 10                                   Aechmea apocalyptica Reitz
#>                          geometry
#> 1  POLYGON ((-41.478 -12.1831,...
#> 2  POLYGON ((-78.1828 -3.26, -...
#> 3  POLYGON ((-96.4525 16.1917,...
#> 4  POLYGON ((-40.4689 -18.5706...
#> 5  POLYGON ((-44.1391 -22.0335...
#> 6  POLYGON ((-41 -20.7089, -41...
#> 7  POLYGON ((-39.16467 -17.048...
#> 8  POLYGON ((-38.92074 -15.997...
#> 9  POLYGON ((-76.0641 0.5833, ...
#> 10 MULTIPOLYGON (((-48.5357 -2...

get_range(range_size = c(1000, 10000))

#> Simple feature collection with 1757 features and 1 field
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: -110.31 ymin: -40.30995 xmax: -34.79361 ymax: 29.9247
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>                                               tax_accepted_name
#> 1          Acanthostachys pitcairnioides (Mez) Rauh & Barthlott
#> 2  Aechmea aenigmatica Lopez-Ferrari, Espejo, Ceja et A.Mendoza
#> 3                        Aechmea aguadocensis Leme & L.Kollmann
#> 4                                     Aechmea aiuruocensis Leme
#> 5                                    Aechmea alegrensis W.Weber
#> 6                         Aechmea amicorum B.R.Silva & H.Luther
#> 7                                         Aechmea ampla L.B.Sm.
#> 8                                       Aechmea anomala L.B.Sm.
#> 9                                    Aechmea apocalyptica Reitz
#> 10                                     Aechmea araneosa L.B.Sm.
#>                          geometry
#> 1  POLYGON ((-41.478 -12.1831,...
#> 2  POLYGON ((-96.4525 16.1917,...
#> 3  POLYGON ((-40.4689 -18.5706...
#> 4  POLYGON ((-44.1391 -22.0335...
#> 5  POLYGON ((-41 -20.7089, -41...
#> 6  POLYGON ((-39.16467 -17.048...
#> 7  POLYGON ((-38.92074 -15.997...
#> 8  POLYGON ((-76.0641 0.5833, ...
#> 9  MULTIPOLYGON (((-48.5357 -2...
#> 10 POLYGON ((-41.2306 -21.3495...